requirements.txt
Django==1.6.5
django-cors-headers
PyYAML
python-ldap
django-auth-ldap
uwsgi
